# backend/internal/api/openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "tela REST API",
    "version": "2026-09-28",
    "summary": "HTTP API of tela, a self-hostable markdown-native team wiki.",
...
    "contact": {
      "name": "tela",
      "url": "https://telawiki.com/support",
      "email": "tela@telawiki.com"
    },
    "license": {
      "name": "AGPL-3.0-or-later",
      "url": "https://github.com/zcag/tela/blob/main/LICENSE"
    }
  },
...
  "servers": [
    {
      "url": "https://telawiki.com",
      "description": "The instance serving this document. A self-hosted instance rewrites this to its own origin."
    }
  ],
  "security": [{ "bearerAuth": [] }, { "sessionCookie": [] }],
...
  "paths": {
...
    "/api/spaces/{id}/files": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["spaces", "attachments"],
        "operationId": "listSpaceFiles",
        "summary": "Every file in a space",
        "description": "Wider than the per-page listing: it also reaches files parented to the space root, which no page lists — the ones a sync or import drops beside the tree. Capped at 500 rows.",
        "parameters": [
          {
            "name": "parent_page_id",
            "in": "query",
            "required": false,
            "description": "A page id to narrow to that page's files, or the literal `root` for files at the space root. Omit for the whole space.",
            "schema": { "type": "string" },
            "examples": { "one page": { "value": "1234" }, "space root": { "value": "root" } }
          }
        ],
        "responses": {
          "200": {
            "description": "Files. `truncated` is true when the 500-row cap was hit.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["files", "truncated"],
                  "properties": {
                    "files": { "type": "array", "items": { "$ref": "#/components/schemas/SpaceFile" } },
                    "truncated": { "type": "boolean" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/BadRequest" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/pages": {
      "get": {
        "tags": ["pages"],
        "operationId": "listPages",
        "summary": "Pages in a space",
        "description": "Flat by default (one level, filtered by `parent_id`), or the whole nested tree with `tree=1`. Bodies are **not** included — use `GET /api/pages/{id}` or `/api/pages/bodies`.",
        "parameters": [
          { "name": "space_id", "in": "query", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
          {
            "name": "parent_id",
            "in": "query",
            "required": false,
            "description": "Direct children of this page. Omit, or pass the literal `null`, for top-level pages. Ignored when `tree=1`.",
            "schema": { "type": "string" }
          },
          {
            "name": "tree",
            "in": "query",
            "required": false,
            "description": "`1` returns the whole space as a nested tree; each node gains a `children` array.",
            "schema": { "type": "string", "enum": ["1"] }
          }
        ],
        "responses": {
          "200": {
            "description": "Pages. Nodes carry `children` only in tree mode.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["pages"],
                  "properties": {
                    "pages": { "type": "array", "items": { "$ref": "#/components/schemas/PageNode" } }
                  }
                }
              }
            }
          },
          "400": {
            "description": "`invalid_query` — `space_id` missing or malformed, or a bad `parent_id`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "404": {
            "description": "`space_not_found`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      },
      "post": {
        "tags": ["pages"],
        "operationId": "createPage",
        "summary": "Create a page",
        "description": "`editor+` on the target space. Two server-side transforms to know about: leading YAML frontmatter in `body` is lifted into `props` (except on deck and sheet pages, whose bodies are stored verbatim), and a leading `# Heading` identical to `title` is stripped.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "required": ["space_id", "title"],
                "properties": {
                  "space_id": { "type": "integer", "format": "int64", "minimum": 1 },
                  "parent_id": { "type": ["integer", "null"], "format": "int64", "description": "Omit or null for a top-level page. Must live in the same space." },
                  "title": { "type": "string", "minLength": 1, "maxLength": 500 },
                  "body": { "type": "string", "description": "Markdown." },
                  "props": { "$ref": "#/components/schemas/PageProps" },
                  "status": { "type": "string", "enum": ["draft", "published"], "description": "A `draft` is labelled in the app and never served on public spaces or share links until published. Defaults to `published`." }
                }
              }
            }
          }
        },
        "responses": {
          "201": { "$ref": "#/components/responses/PageEnvelope" },
          "400": {
            "description": "`invalid_json`, `invalid_title`, `invalid_space_id`, `invalid_parent_id`, `space_not_found`, `parent_not_found`, or `parent_space_mismatch`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "402": { "$ref": "#/components/responses/QuotaExceeded" },
          "403": { "$ref": "#/components/responses/Forbidden" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/pages/{id}": {
      "parameters": [{ "$ref": "#/components/parameters/PathId" }],
      "get": {
        "tags": ["pages"],
        "operationId": "getPage",
        "summary": "Get a page",
        "description": "Returns the canonical markdown body plus how the page is exposed to the outside world.",
        "responses": {
          "200": {
            "description": "The page.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["page", "exposure"],
                  "properties": {
                    "page": { "$ref": "#/components/schemas/Page" },
                    "exposure": { "$ref": "#/components/schemas/Exposure" }
                  }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "403": { "$ref": "#/components/responses/ForbiddenOrMissingPage" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      },
      "patch": {
        "tags": ["pages"],
        "operationId": "updatePage",
        "summary": "Update a page's title, body, properties, or draft status",
        "description": "`editor+`. A change to `title` or `body` snapshots a revision first. **`props` replaces the whole bag** — send the properties you want to keep. There is no section-level patch over REST; the `patch_page` MCP tool is the only place heading-path targeting exists.",
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "type": "object",
                "minProperties": 1,
                "properties": {
                  "title": { "type": "string", "minLength": 1, "maxLength": 500 },
                  "body": { "type": "string", "description": "The whole markdown document." },
                  "props": { "$ref": "#/components/schemas/PageProps" },
                  "status": { "type": "string", "enum": ["draft", "published"], "description": "A `draft` is labelled in the app and never served on public spaces or share links until published." },
                  "base_version": { "type": "integer", "format": "int64", "description": "The page `version` you last read. If the page has changed since, the write is refused with `409 version_conflict` instead of overwriting that edit. Omit for last-write-wins." }
                }
              }
            }
          }
        },
        "responses": {
          "200": { "$ref": "#/components/responses/PageEnvelope" },
          "400": {
            "description": "`invalid_json`, `no_fields`, `invalid_title`, or `invalid_status`.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "403": { "$ref": "#/components/responses/ForbiddenOrMissingPage" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "409": {
            "description": "`version_conflict`: `base_version` is stale. Re-read the page and reapply your change.",
            "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
          },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      },
      "delete": {
        "tags": ["pages"],
        "operationId": "deletePage",
        "summary": "Delete a page and its subtree",
        "description": "`editor+`. Soft delete — the page and **all of its descendants** move to the space trash.",
        "responses": {
          "204": { "description": "Deleted." },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "403": { "$ref": "#/components/responses/ForbiddenOrMissingPage" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/files/{space_id}/{file}": {
      "parameters": [
        { "name": "space_id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } },
        { "name": "file", "in": "path", "required": true, "description": "The 64-character lower-case SHA-256 of the content, optionally with a cosmetic extension (`<hash>.png`). Take it from an attachment's `url`.", "schema": { "type": "string", "pattern": "^[0-9a-f]{64}(\\.[A-Za-z0-9]+)?$" } }
      ],
      "get": {
        "tags": ["attachments"],
        "operationId": "getFileBytes",
        "summary": "Fetch stored file bytes",
        "description": "**Public and unauthenticated** — knowing the content hash is the authorization, which is why hashes are not enumerable and error responses never distinguish \"absent\" from \"not yours\". Immutable and cacheable for a year; supports `If-None-Match`. Only common raster image types are served `inline`; everything else, SVG included, is `attachment`.",
        "security": [],
        "responses": {
          "200": {
            "description": "The bytes.",
            "headers": {
              "ETag": { "schema": { "type": "string" } },
              "Cache-Control": { "description": "`public, max-age=31536000, immutable`", "schema": { "type": "string" } },
              "Content-Disposition": { "schema": { "type": "string" } }
            },
            "content": { "*/*": { "schema": { "type": "string", "format": "binary" } } }
          },
          "304": { "description": "Not modified." },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/share/{token}/page/{page_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/ShareToken" },
        { "name": "page_id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } }
      ],
      "get": {
        "tags": ["shares"],
        "operationId": "getPublicSharePage",
        "summary": "Read a descendant page of a share",
        "description": "Only for shares created with `include_descendants`. A page outside the shared subtree answers 404 exactly like a missing one.",
        "security": [],
        "responses": {
          "200": {
            "description": "The page.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["page"],
                  "properties": { "page": { "$ref": "#/components/schemas/SharePage" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "401": { "$ref": "#/components/responses/PasswordRequired" },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/public/spaces/{id}/pages/{page_id}": {
      "parameters": [
        { "$ref": "#/components/parameters/PathId" },
        { "name": "page_id", "in": "path", "required": true, "schema": { "type": "integer", "format": "int64", "minimum": 1 } }
      ],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicSpacePage",
        "summary": "Read a page in a published space",
        "description": "No login. Publishing grants read only — there is no public write path anywhere under `/api/public/`.",
        "security": [],
        "responses": {
          "200": {
            "description": "The page.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["page"],
                  "properties": { "page": { "$ref": "#/components/schemas/PublicPage" } }
                }
              }
            }
          },
          "400": { "$ref": "#/components/responses/InvalidId" },
          "404": { "$ref": "#/components/responses/NoSuchPublicSpace" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
    "/api/public/files/{hash}": {
      "parameters": [
        { "name": "hash", "in": "path", "required": true, "description": "The file's content hash, full or the 12-character short form used in `/f/` links.", "schema": { "type": "string" } }
      ],
      "get": {
        "tags": ["public"],
        "operationId": "getPublicFileMeta",
        "summary": "Metadata for a shareable file link",
        "description": "Backs the human-facing `/f/{hash}/{name}` page. The owning `page` is included only when its space is published.",
        "security": [],
        "responses": {
          "200": {
            "description": "File metadata.",
            "content": {
              "application/json": {
                "schema": {
                  "type": "object",
                  "required": ["file"],
                  "properties": { "file": { "$ref": "#/components/schemas/PublicFile" } }
                }
              }
            }
          },
          "404": { "$ref": "#/components/responses/NotFound" },
          "500": { "$ref": "#/components/responses/Internal" }
        }
      }
    },
...
  },
  "components": {
    "securitySchemes": {
      "bearerAuth": {
        "type": "http",
        "scheme": "bearer",
        "description": "A personal access token: `Authorization: Bearer tela_pat_<43 chars>`. Scope `read` allows GET and HEAD only; `write` and `admin` allow any method. A token may be pinned to a single space."
      },
      "sessionCookie": {
        "type": "apiKey",
        "in": "cookie",
        "name": "tela_session",
        "description": "The browser session cookie set by `POST /api/auth/login`. HttpOnly, SameSite=Lax, 30 days."
      }
    },
    "parameters": {
      "PathId": {
        "name": "id",
        "in": "path",
        "required": true,
        "schema": { "type": "integer", "format": "int64", "minimum": 1 }
      },
      "ShareToken": {
        "name": "token",
        "in": "path",
        "required": true,
        "description": "The unguessable share token from a share link's `url`.",
        "schema": { "type": "string" }
      },
...
    },
    "responses": {
...
      "PageEnvelope": {
        "description": "The page.",
        "content": {
          "application/json": {
            "schema": {
              "type": "object",
              "required": ["page"],
              "properties": { "page": { "$ref": "#/components/schemas/Page" } }
            }
          }
        }
      },
...
      "BadRequest": {
        "description": "Malformed request. `code` is usually `bad_request` or `invalid_json`.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "InvalidId": {
        "description": "`invalid_id` — a path parameter was not a positive integer.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "Forbidden": {
        "description": "`forbidden` (no access, or the role is too low), `viewer_no_write`, `api_key_scope` (the token's scope forbids this method), or `api_key_space_scope` (the token is pinned to another space).",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "ForbiddenOrMissingPage": {
        "description": "`forbidden` — **also returned when the page does not exist**, so page ids cannot be enumerated. When the page happens to live in a published space, `code` is `forbidden_public` and `public_path` points at the no-login reader. Also `api_key_scope` / `api_key_space_scope`.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NotFound": {
        "description": "`not_found`.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "NoSuchPublicSpace": {
        "description": "`not_found` — no such published space. Private and non-existent are indistinguishable.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "PasswordRequired": {
        "description": "`password_required` — this share is password-gated; call `POST /api/share/{token}/auth` first.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
      "QuotaExceeded": {
        "description": "`quota_exceeded` — a plan limit (pages, storage) was reached.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      },
...
      "Internal": {
        "description": "`internal` — server error.",
        "content": { "application/json": { "schema": { "$ref": "#/components/schemas/Error" } } }
      }
    },
    "schemas": {
      "Error": {
        "type": "object",
        "description": "The error envelope for every JSON route. Branch on `code`, not on the message.",
        "required": ["error", "code"],
        "properties": {
          "error": { "type": "string", "description": "Human-readable message. Not stable — do not parse it." },
          "code": { "type": "string", "description": "Machine-readable code.", "examples": ["forbidden", "not_found", "api_key_scope"] },
          "public_path": { "type": "string", "description": "Only with `code: forbidden_public` — where to read this page without an account." }
        }
      },
      "Timestamp": {
        "type": "string",
        "description": "`YYYY-MM-DD HH:MM:SS` in UTC.",
        "examples": ["2026-09-09 14:03:21"]
      },
...
      "PageProps": {
        "type": "object",
        "additionalProperties": true,
        "description": "Free-form page properties. Reserved keys are dropped. Two are load-bearing: `deck: true` makes the body a Slidev presentation, `sheet: true` makes it a spreadsheet — both are then stored verbatim, frontmatter included.",
        "examples": [{ "status": "draft", "owner": "platform-team" }]
      },
      "Page": {
        "type": "object",
        "required": ["id", "space_id", "parent_id", "title", "body", "position", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "space_id": { "type": "integer", "format": "int64" },
          "parent_id": { "type": ["integer", "null"], "format": "int64" },
          "title": { "type": "string" },
          "body": { "type": "string", "description": "Canonical markdown. This is the whole document — tela has no block model." },
          "position": { "type": "integer", "format": "int64", "description": "Sort order among siblings." },
          "props": { "$ref": "#/components/schemas/PageProps" },
          "created_at": { "$ref": "#/components/schemas/Timestamp" },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "version": { "type": "integer", "format": "int64", "description": "Advances on every title/body/props change. Send it back as `base_version` on update to refuse overwriting an edit you haven't seen." },
          "status": { "type": "string", "enum": ["draft", "published"], "description": "A `draft` is labelled in the app and never served on public spaces or share links until published." }
        }
      },
      "PageNode": {
        "allOf": [
          { "$ref": "#/components/schemas/Page" },
          {
            "type": "object",
            "required": ["exposure"],
            "properties": {
              "exposure": { "$ref": "#/components/schemas/Exposure" },
              "children": { "type": "array", "items": { "$ref": "#/components/schemas/PageNode" }, "description": "Present only in tree mode (`?tree=1`)." }
            }
          }
        ]
      },
      "Exposure": {
        "type": "object",
        "description": "How far outside the space this page is readable.",
        "required": ["state", "inherited", "expires_at"],
        "properties": {
          "state": { "type": "string", "enum": ["private", "public", "password"], "description": "`public` = reachable with no credential; `password` = reachable with a share password." },
          "inherited": { "type": "boolean", "description": "True when the exposure comes from an ancestor's subtree share rather than this page." },
          "expires_at": { "type": ["string", "null"] }
        }
      },
...
      "SpaceFile": {
        "allOf": [
          { "$ref": "#/components/schemas/Attachment" },
          {
            "type": "object",
            "properties": {
              "parent_page_id": { "type": "integer", "format": "int64", "description": "Absent when the file sits at the space root." },
              "parent_title": { "type": "string", "description": "Absent when the file sits at the space root." }
            }
          }
        ]
      },
...
      "SharePage": {
        "type": "object",
        "required": ["id", "title", "body", "updated_at"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "body": { "type": "string", "description": "Canonical markdown." },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" }
        }
      },
...
      "PublicPage": {
        "type": "object",
        "required": ["id", "title", "body", "created_at", "updated_at"],
        "properties": {
          "id": { "type": "integer", "format": "int64" },
          "title": { "type": "string" },
          "body": { "type": "string", "description": "Canonical markdown." },
          "props": { "$ref": "#/components/schemas/PageProps" },
          "deck": {
            "type": "object",
            "description": "Present only on a presentation page.",
            "properties": {
              "present_path": { "type": "string" },
              "cover_path": { "type": "string" }
            }
          },
          "created_at": { "$ref": "#/components/schemas/Timestamp" },
          "updated_at": { "$ref": "#/components/schemas/Timestamp" },
          "author": { "type": "string" },
          "editor": { "type": "string" }
        }
      },
...
      "PublicFile": {
        "type": "object",
        "required": ["hash", "short", "name", "mime", "kind", "byte_size", "url", "path"],
        "properties": {
          "hash": { "type": "string" },
          "short": { "type": "string", "description": "The 12-character hash prefix used in `/f/` links." },
          "name": { "type": "string" },
          "mime": { "type": "string" },
          "kind": { "type": "string", "description": "Coarse category driving how the file page previews it." },
          "byte_size": { "type": "integer", "format": "int64" },
          "url": { "type": "string", "description": "Where the bytes are served." },
          "path": { "type": "string", "description": "The file page's own path." },
          "page": {
            "type": "object",
            "description": "The page this file hangs off. Present only when that page's space is published.",
            "properties": {
              "id": { "type": "integer", "format": "int64" },
              "title": { "type": "string" },
              "path": { "type": "string" },
              "space_name": { "type": "string" }
            }
          }
        }
      },
...
    }
  }
}